# api/openapi.yaml
openapi: 3.0.0
info:
  title: Murder Mystery API
  description: API for retrieving murder mystery game information
  version: 1.0.0
servers:
  - url: https://api.murder-mystery-fans.com
paths:
  /health:
    get:
      summary: Check the health of the server
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    description: The message of the server
                    example: "ok"

  /murder_mysteries:
    get:
      summary: Get a list of murder mysteries
      description: Retrieve murder mysteries.
      parameters:
        - name: yearMonth
          in: query
          description: The year and month for which to retrieve games (YYYY-MM format)
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
      responses:
        "200":
          description: A list of murder mysteries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MurderMystery'
        "400":
          description: Invalid month format
        "500":
          description: Internal server error

  /murder_mysteries/{id}:
    get:
      summary: Get details of a specific murder mystery
      description: Retrieve all details of a specific murder mystery by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the murder mystery game
          schema:
            type: integer
      responses:
        "200":
          description: Detailed information of the murder mystery game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MurderMysteryDetail'
        "400":
          description: Invalid ID format
        "404":
          description: Murder mystery not found
        "500":
          description: Internal server error

components:
  schemas:
    MurderMystery:
      type: object
      required:
        - id
        - title
        - release_date
      properties:
        id:
          type: integer
          description: Unique identifier for the game
        title:
          type: string
          description: Title of the game
        image_link:
          type: string
          format: uri
          description: URL of the game's image
        amazon_link:
          type: string
          format: uri
          description: Amazon purchase link
        official_link:
          type: string
          format: uri
          description: Official website link
        release_date:
          type: string
          format: date
          description: Release date of the game
    MurderMysteryDetail:
        type: object
        properties:
          id:
            type: integer
          title:
            type: string
          release_date:
            type: string
            format: date
          platform:
            type: string
          language:
            type: string
          official_link:
            type: string
            format: uri
          amazon_link:
            type: string
            format: uri
          surugaya_link:
            type: string
            format: uri
          steam_link:
            type: string
            format: uri
          epicgames_link:
            type: string
            format: uri
          dlsite_link:
            type: string
            format: uri
          rakuten_books_link:
            type: string
            format: uri
          bookwalker_link:
            type: string
            format: uri
          fanza_games_link:
            type: string
            format: uri
          price:
            type: number
            format: float
          discount:
            type: number
            format: float
          original_price:
            type: number
            format: float
          image_link:
            type: string
            format: uri
          description:
            type: string
          players_min:
            type: integer
          players_max:
            type: integer
          play_time:
            type: integer
          requires_gm:
            type: boolean
          online_supported:
            type: boolean
          genre:
            type: string
          replayability:
            type: string
          age_recommendation:
            type: string